4bit, 16bit 정수 입력 처리

+++ CS23/STEP_1_DIGITAL_LOGIC_CIRCUIT/STEP_1-1_NAND_GATE_SWIFT/STEP_1-1_NAND_GATE_SWIFT/InputView.swift
@@ -17,11 +17,11 @@ struct InputView {
         return input
     }
     
-    static func convertDecimalToBinaryNumber(_ input: String) -> UInt8{
-        return UInt8(input) ?? 0b0
+    static func convertDecimalToBinaryNumber(_ input: String) -> UInt64{
+        return UInt64(input) ?? 0b0
     }
     
-    static func readDecimalNumber () -> UInt8 {
+    static func readDecimalNumber () -> UInt64 {
         let input = readNumber("정수 입력 > ")
         return convertDecimalToBinaryNumber(input)
     }
